Move api urls and views to chamber of deputies and core
