+++ jarbas/chamber_of_deputies/urls.py
@@ -1,12 +1,11 @@
 from django.conf.urls import url
 
-from jarbas.api.views import (
-    ApplicantListView,
-    CompanyDetailView,
-    ReceiptDetailView,
-    ReimbursementDetailView,
+from jarbas.chamber_of_deputies.views import (
     ReimbursementListView,
+    ReimbursementDetailView,
+    ReceiptDetailView,
     SameDayReimbursementListView,
+    ApplicantListView,
     SubquotaListView,
 )
 
@@ -31,11 +30,6 @@ urlpatterns = [
         r'^chamber_of_deputies/reimbursement/(?P<document_id>\d+)/same_day/$',
         SameDayReimbursementListView.as_view(),
         name='reimbursement-same-day'
-    ),
-    url(
-        r'^company/(?P<cnpj>\d{14})/$',
-        CompanyDetailView.as_view(),
-        name='company-detail'
     ),
     url(r'^chamber_of_deputies/applicant/$', ApplicantListView.as_view(), name='applicant-list'),
     url(r'^chamber_of_deputies/subquota/$', SubquotaListView.as_view(), name='subquota-list')
